requirements.txt
discord.py==2.4.0
PyNaCl==1.5.0
yt_dlp==2024.5.27
